refactor(era): Isolate `EraStream` in a dedicated module
