+++ crates/era-downloader/src/client.rs
@@ -1,13 +1,6 @@
-use futures_util::stream::{FuturesOrdered, Stream, StreamExt};
-use pin_project::pin_project;
+use futures_util::stream::StreamExt;
 use reqwest::{Client, IntoUrl, Url};
-use std::{
-    future::Future,
-    path::Path,
-    pin::Pin,
-    str::FromStr,
-    task::{Context, Poll},
-};
+use std::{path::Path, str::FromStr};
 use tokio::{
     fs::{self, File},
     io::{self, AsyncBufReadExt, AsyncWriteExt},
@@ -27,9 +20,7 @@ impl EraClient {
     pub fn new(client: Client, url: Url, folder: Box<Path>) -> Self {
         Self { client, url, folder }
     }
-}
 
-impl EraClient {
     /// Performs a GET request on `url` and stores the response body into a file located within
     /// the `folder`.
     pub async fn download_to_file(&mut self, url: impl IntoUrl) -> eyre::Result<Box<Path>> {
@@ -93,7 +84,7 @@ impl EraClient {
 
     /// Fetches the list of ERA1 files from `url` and stores it in a file located within `folder`.
     pub async fn fetch_file_list(&self) -> eyre::Result<()> {
-        let response = self.client.get(self.url.clone()).send().await.unwrap();
+        let response = self.client.get(self.url.clone()).send().await?;
 
         let mut stream = response.bytes_stream();
         let path = self.folder.to_path_buf().join("index.html");
@@ -140,131 +131,6 @@ impl EraClient {
 
     fn file_name_to_number(&self, file_name: &str) -> Option<u64> {
         file_name.split('-').nth(1).and_then(|v| u64::from_str(v).ok())
-    }
-}
-
-struct EraStream {
-    pub download_stream: DownloadStream,
-    pub starting_stream: StartingStream,
-}
-
-impl Stream for EraStream {
-    type Item = eyre::Result<Box<Path>>;
-
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        if let Poll::Ready(Some(fut)) = Pin::new(&mut self.starting_stream).poll_next(cx) {
-            self.download_stream.downloads.push_back(fut);
-        }
-
-        let downloading = self.download_stream.downloads.len();
-        self.starting_stream.downloading(downloading);
-
-        Pin::new(&mut self.download_stream).poll_next(cx)
-    }
-}
-
-type DownloadFuture = Pin<Box<dyn Future<Output = eyre::Result<Box<Path>>>>>;
-
-#[pin_project]
-struct DownloadStream {
-    #[pin]
-    pub downloads: FuturesOrdered<DownloadFuture>,
-}
-
-impl Stream for DownloadStream {
-    type Item = eyre::Result<Box<Path>>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let project = self.project();
-        project.downloads.poll_next(cx)
-    }
-}
-
-#[pin_project]
-struct StartingStream {
-    client: EraClient,
-    downloading: usize,
-    files_count: Pin<Box<dyn Future<Output = usize> + Send + Sync + 'static>>,
-    next_url: Pin<Box<dyn Future<Output = Option<Url>> + Send + Sync + 'static>>,
-    files_count_ready: bool,
-    state: State,
-    max_files: usize,
-    max_concurrent_downloads: usize,
-}
-
-#[derive(Debug, PartialEq)]
-enum State {
-    Initial,
-    Missing(usize),
-    NextUrl(usize),
-}
-
-impl StartingStream {
-    fn downloading(&mut self, downloading: usize) {
-        self.downloading = downloading;
-
-        if !self.files_count_ready {
-            let client = self.client.clone();
-
-            Pin::new(&mut self.files_count)
-                .set(Box::pin(async move { client.files_count().await }));
-            self.files_count_ready = true;
-        }
-    }
-}
-
-impl Stream for StartingStream {
-    type Item = Pin<Box<dyn Future<Output = eyre::Result<Box<Path>>>>>;
-
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let downloading = self.downloading;
-
-        if self.state == State::Initial && self.files_count_ready {
-            match Pin::new(&mut self.files_count).poll(cx) {
-                Poll::Ready(downloaded) => {
-                    let max_missing = (downloaded + downloading)
-                        .saturating_sub(self.max_files)
-                        .max(self.max_concurrent_downloads);
-
-                    self.files_count_ready = false;
-                    self.state = State::Missing(max_missing);
-                }
-                Poll::Pending => return Poll::Pending,
-            }
-        }
-
-        if let State::Missing(max_missing) = self.state {
-            if max_missing > 0 {
-                let client = self.client.clone();
-
-                Pin::new(&mut self.next_url).set(Box::pin(async move { client.next_url().await }));
-
-                self.state = State::NextUrl(max_missing);
-            } else {
-                self.state = State::Initial;
-            }
-        }
-
-        if let State::NextUrl(max_missing) = self.state {
-            return match Pin::new(&mut self.next_url).poll(cx) {
-                Poll::Ready(url) => {
-                    self.state = State::Missing(max_missing.saturating_sub(1));
-
-                    if let Some(url) = url {
-                        let mut client = self.client.clone();
-
-                        Poll::Ready(Some(Box::pin(
-                            async move { client.download_to_file(url).await },
-                        )))
-                    } else {
-                        Poll::Ready(None)
-                    }
-                }
-                Poll::Pending => Poll::Pending,
-            }
-        }
-
-        Poll::Pending
     }
 }
