+++ crates/era-downloader/src/lib.rs
@@ -1,4 +1,6 @@
-//! A streamable interface for downloading ERA1 files.
+//! An asynchronous stream interface for downloading ERA1 files.
 mod client;
+mod stream;
 
 pub use client::EraClient;
+pub use stream::EraStream;
